+++ NEWS.md
@@ -1,8 +1,12 @@
 # argon2 (development version)
+#### New features
+* Added `char_as_raw` to convert a string of hexadecimal encoded values to raw.
+
 
 # argon2 0.7.1
 #### Miscellaneous
 * Trying to fix compilation issues in Windows.
+
 
 # argon2 0.7.0
 #### New features